+++ AlchemistCode/Cards/Common/Poultice.cs
@@ -9,7 +9,7 @@ namespace Alchemist.AlchemistCode.Cards.Common;
 
 public class Poultice : AlchemistCard
 {
-    public Poultice() : base(1, CardType.Skill, CardRarity.Common, TargetType.Self)
+    public Poultice() : base(0, CardType.Skill, CardRarity.Common, TargetType.Self)
     {
         WithPower<RegenPower>(2, 1);
     }